add settings required by Bitbucket client credentials grant
To use this, you must first create an OAuth consumer in Bitbucket to get the
`client_id` and `client_secret` values to make your token request using the
client credentials grant. The `/user` endpoint is used to validate the
datasource settings, so the consumer needs to have at least the `account`
scope selected.

Bitbucket does the client credentials grant a little bit differently
compared to the example in the Grafana docs because it uses the `client_id`
and `client_secret` as the username and password portion of HTTP Basic auth
instead of having them as request data. ...although I did see that Bitbucket
is supported as a type of "generic oauth" provider for logins, so it might
support using them as form data too. I was still figuring stuff out when
I initially tried using them as form data, so I'll need to retest to
confirm if both usages are acceptable.

Later down the line I would like to also add support the ability to
forward OAuth identity for the logged-in user. This could add some
security since a user will only be able to see data for projects they have
been given access too. I don't want to only support that though since I do not
want to limit the use of other providers for loggin in to Grafana.

* [Create Bitbucket OAuth Consumer](https://support.atlassian.com/bitbucket-cloud/docs/use-oauth-on-bitbucket-cloud/#Create-a-consumer)
* [Bitbucket Client Credentials Grand](https://developer.atlassian.com/cloud/bitbucket/oauth-2/#4--client-credentials-grant--4-4-)
* [Add authorization for Data Source Plugins](https://grafana.com/docs/grafana/latest/developers/plugins/add-authentication-for-data-source-plugins)
* [Forward OAuth Identity](https://grafana.com/docs/grafana/latest/developers/plugins/add-authentication-for-data-source-plugins/#forward-oauth-identity-for-the-logged-in-user)