test querying datasource
Added some of the fields from the GitHub datasource QueryEditor to test
generating data frames from API data.

I chose to use commits for the test data.

Some early observations:
The goal is to make the Bitbucket frames compatible
with GitHub so that they could be easily combined, but the data structures
are completely different. (I don't know what I was expecting) So unfortunately
people will have to do some transforms.

Another thing is that the datasource generator uses "LegacyForms", which
I assume is an older set of components, and there is no documentation for
those. The GitHub plugin uses a set of functional React components that are
also different from the documented components.